requirements.txt
# Nalarbanjir v2 — core dependencies (no optional ML packages)
# Install with:  pip install -r requirements.txt

# ── Numerics ──────────────────────────────────────────────────────────────
numpy>=1.26.0

# ── Web framework ─────────────────────────────────────────────────────────
fastapi>=0.135.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
python-multipart>=0.0.9
httpx>=0.27.0         # used by FastAPI TestClient

# ── Utilities ─────────────────────────────────────────────────────────────
pyyaml>=6.0
python-json-logger>=2.0.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── Optional — ML (requires separate install or Docker layer) ─────────────
# torch>=2.2.0         # uncomment for GPU-accelerated FloodNet training
# scipy>=1.12.0        # uncomment for advanced interpolation
# scikit-learn>=1.4.0  # uncomment for random forest baseline
